UHF-9708: Unify global and local blocks
