+++ modules/helfi_etusivu_entities/helfi_etusivu_entities.info.yml
@@ -5,3 +5,5 @@ dependencies:
   - external_entities:external_entities
   - helfi_api_base:helfi_api_base
   - helfi_platform_config:helfi_platform_config
+  - helfi_platform_config:helfi_node_announcement
+  - helfi_platform_config:helfi_node_survey
