+++ modules/helfi_etusivu_entities/helfi_etusivu_entities.install
@@ -18,28 +18,13 @@ function helfi_etusivu_entities_install($is_syncing) : void {
   }
 
   helfi_etusivu_entities_grant_permissions();
-  _helfi_etusivu_entities_create_block();
+  helfi_etusivu_entities_create_blocks();
 }
 
 /**
- * Grants required permissions.
+ * Create required blocks.
  */
-function helfi_etusivu_entities_grant_permissions() : void {
-  $permissions = [
-    'view helfi_announcements external entity',
-    'view helfi_surveys external entity',
-  ];
-
-  helfi_platform_config_grant_permissions([
-    'anonymous' => $permissions,
-    'authenticated' => $permissions,
-  ]);
-}
-
-/**
- * Added external entity for announcements and enable block.
- */
-function _helfi_etusivu_entities_create_block() : void {
+function helfi_etusivu_entities_create_blocks(): void {
   /** @var Drupal\helfi_platform_config\Helper\BlockInstaller $block_installer */
   $block_installer = Drupal::service('helfi_platform_config.helper.block_installer');
 
@@ -55,4 +40,36 @@ function _helfi_etusivu_entities_create_block() : void {
     ['block' => $block, 'variations' => $variations] = $block_config;
     $block_installer->install($block, $variations);
   }
+}
+
+/**
+ * Implements hook_themes_installed().
+ */
+function helfi_etusivu_entities_themes_installed($theme_list) : void {
+  /** @var Drupal\helfi_platform_config\Helper\BlockInstaller $block_installer */
+  $block_installer = Drupal::service('helfi_platform_config.helper.block_installer');
+
+  foreach ($theme_list as $theme) {
+    if (in_array($theme, ['stark', 'hdbt', 'hdbt_subtheme'])) {
+      foreach (helfi_etusivu_entities_get_block_configurations($theme) as $block_config) {
+        ['block' => $block, 'variations' => $variations] = $block_config;
+        $block_installer->install($block, $variations);
+      }
+    }
+  }
+}
+
+/**
+ * Grants required permissions.
+ */
+function helfi_etusivu_entities_grant_permissions() : void {
+  $permissions = [
+    'view helfi_announcements external entity',
+    'view helfi_surveys external entity',
+  ];
+
+  helfi_platform_config_grant_permissions([
+    'anonymous' => $permissions,
+    'authenticated' => $permissions,
+  ]);
 }
