+++ modules/helfi_etusivu_entities/helfi_etusivu_entities.module
@@ -7,35 +7,46 @@
 
 declare(strict_types=1);
 
-use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\Entity\EntityTypeInterface;
 use Drupal\Core\StringTranslation\TranslatableMarkup;
 use Drupal\entity\BundleFieldDefinition;
-use Drupal\helfi_api_base\Cache\CacheTagInvalidatorInterface;
-use Drupal\helfi_api_base\Environment\EnvironmentResolverInterface;
-use Drupal\helfi_api_base\Environment\Project;
-use Drupal\helfi_etusivu_entities\Plugin\ExternalEntities\StorageClient\Announcements;
-use Drupal\helfi_etusivu_entities\Plugin\ExternalEntities\StorageClient\Surveys;
 
 /**
  * Gets the block configurations.
+ *
+ * Example block:
+ *
+ * @code
+ * [
+ *   'breadbrumbs' => [
+ *     'block' => [
+ *       ...
+ *     ],
+ *     'variations' => [
+ *       ...
+ *     ],
+ *   ],
+ * ];
+ * @endcode
  *
  * @return array[]
  *   The block configurations.
  */
 function helfi_etusivu_entities_get_block_configurations(string $theme) : array {
   return [
-    'global_announcement' => [
+    // External entity version of this block is
+    // enabled by helfi_etusivu_entities module.
+    'announcements' => [
       'block' => [
-        'id' => 'globalannouncement',
-        'plugin' => 'global_announcements',
+        'id' => 'announcements',
+        'plugin' => 'announcements',
         'settings' => [
-          'id' => 'global_announcements',
-          'label' => 'Global announcement',
-          'provider' => 'helfi_etusivu_entities',
+          'id' => 'announcements',
+          'label' => 'Announcements',
+          'provider' => 'helfi_node_announcement',
         ],
-        'provider' => 'helfi_etusivu_entities',
-        'weight' => -14,
+        'provider' => 'helfi_node_announcement',
+        'weight' => -15,
       ],
       'variations' => [
         [
@@ -48,17 +59,19 @@ function helfi_etusivu_entities_get_block_configurations(string $theme) : array 
         ],
       ],
     ],
-    'global_survey' => [
+    // External entity version of this block is
+    // enabled by helfi_etusivu_entities module.
+    'surveys' => [
       'block' => [
-        'id' => 'globalsurvey',
-        'plugin' => 'global_surveys',
+        'id' => 'surveys',
+        'plugin' => 'surveys',
         'settings' => [
-          'id' => 'global_surveys',
-          'label' => 'Global surveys',
-          'provider' => 'helfi_etusivu_entities',
+          'id' => 'surveys',
+          'label' => 'Surveys',
+          'provider' => 'helfi_node_survey',
         ],
-        'provider' => 'helfi_etusivu_entities',
-        'weight' => -14,
+        'provider' => 'helfi_node_survey',
+        'weight' => -15,
       ],
       'variations' => [
         [
