+++ modules/helfi_node_survey/config/optional/language/fi/field.field.node.survey.field_survey_content_pages.yml
@@ -1,1 +1,2 @@
 label: 'Näytä sisältösivuilla'
+description: 'Näytetään kaikilla sivuilla jos jätetään tyhjäksi.'
