+++ modules/helfi_platform_config_base/helfi_platform_config_base.info.yml
@@ -12,6 +12,7 @@ dependencies:
   - helfi_media_map:helfi_media_map
   - helfi_media_remote_video:helfi_media_remote_video
   - helfi_node_announcement:helfi_node_announcement
+  - helfi_node_survey:helfi_node_survey
   - helfi_node_landing_page:helfi_node_landing_page
   - helfi_node_page:helfi_node_page
   - helfi_paragraphs_accordion:helfi_paragraphs_accordion
