+++ phpstan.neon
@@ -29,9 +29,6 @@ parameters:
       message: '#^Access to an undefined property Drupal\\Core\\Field\\FieldItemInterface::\$alt.#'
       path: modules/helfi_paragraphs_contact_card_listing/src/Entity/ContactCard.php
     -
-      message: '#^Unsafe usage of new static#'
-      path: modules/helfi_node_announcement/src/Plugin/Block/AnnouncementsBlockBase.php
-    -
       message: '#^Access to an undefined property Drupal\\Core\\Field\\FieldItemListInterface<Drupal\\Core\\Field\\FieldItemInterface>::\$iframe_title.#'
       path: modules/helfi_media_remote_video/src/Entity/RemoteVideo.php
     -
